+++ forui_internal_gen/lib/forui_internal_gen.dart
@@ -12,6 +12,7 @@ $defaultFileHeader
 // coverage:ignore-file
 // ignore_for_file: unnecessary_ignore
 // ignore_for_file: avoid_positional_boolean_parameters
+// ignore_for_file: unrelated_type_equality_checks
 ''';
 
 const _designHeader =